tidy: Use size_t for array offsets in mythedid.
When using "uint", all values have to be implicitly converted from a
32-bit number to the CPU register size (which is 64-bit on most
current CPUs).  Using size_t means that the values are already in the
native register size.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone/implicit-widening-of-multiplication-result.html

+++ mythtv/libs/libmythui/mythedid.h
@@ -53,15 +53,15 @@ class MUI_PUBLIC MythEDID
     void        Parse             ();
     bool        ParseBaseBlock    (const quint8* Data);
     void        ParseDisplayDescriptor(const quint8* Data, uint Offset);
-    void        ParseDetailedTimingDescriptor(const quint8* Data, uint Offset);
-    bool        ParseCTA861       (const quint8* Data, uint Offset);
-    bool        ParseCTABlock     (const quint8* Data, uint Offset);
-    bool        ParseVSDB         (const quint8* Data, uint Offset, uint Length);
-    bool        ParseExtended     (const quint8* Data, uint Offset, uint Length);
+    void        ParseDetailedTimingDescriptor(const quint8* Data, size_t Offset);
+    bool        ParseCTA861       (const quint8* Data, size_t Offset);
+    bool        ParseCTABlock     (const quint8* Data, size_t Offset);
+    bool        ParseVSDB         (const quint8* Data, size_t Offset, size_t Length);
+    bool        ParseExtended     (const quint8* Data, size_t Offset, size_t Length);
 
     bool        m_valid           { false };
     QByteArray  m_data            { };
-    uint        m_size            { 0 };
+    size_t      m_size            { 0 };
     quint8      m_minorVersion    { 0 };
     QSize       m_displaySize     { };    // N.B. Either size or aspect are valid
     double      m_displayAspect   { 0.0 };
